// src/app/business-logic/model/events/eventsDebugImagesResponse.ts
/**
 * events
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.14
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {DebugImagesResponseMetrics} from '~/business-logic/model/events/debugImagesResponseMetrics';


export interface EventsDebugImagesResponse {
    /**
     * Debug image events grouped by task metrics and iterations
     */
    metrics?: Array<DebugImagesResponseMetrics>;
    /**
     * Scroll ID for getting more results
     */
    scroll_id?: string;
}